Remove abstract operators from BigNumber, rename divide methods in BigDecimal
It turned out that there was no value in having a generic interface for operations, when every concrete class could really only accept instances of itself, and more importantly, would only return instances of itself. This was also a problem for division, that cannot be handled in the same way for BigInteger, BigDecimal and BigRational.

Methods changed in BigDecimal:

- dividedBy() is now exactlyDividedBy(), and always calculates the scale automatically
- dividedToScale() is now dividedBy(), and does not not make any assumption on the scale any more